% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_testing.R
\name{add_diffcyt_statistics}
\alias{add_diffcyt_statistics}
\title{Add statistical results from the diffcyt package to the fcd.}
\usage{
add_diffcyt_statistics(fcd = condor, input, group1 = "ctrl", group2 = "pat")
}
\arguments{
\item{fcd}{flow cytometry dataset, that has been subjected to clustering or cell type label prediction with cyCONDOR before.}

\item{input}{dataframe containing the statistical results from the diffcyt package. Example: \code{input <- as.data.frame(diffcyt::topTable(res_DA, all = TRUE))}.}

\item{group1}{First group of the comparison.}

\item{group2}{Second group of the comparison.}
}
\value{
the fcd containing the statistical results from the diffcyt package
}
\description{
Wrapper function to include the statistical results calculated using the diffcyt package into the fcd file.
}
\details{
add_diffcyt_statistics
}
