% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{change_param_name}
\alias{change_param_name}
\title{change_param}
\usage{
change_param_name(fcd, old_names, new_names)
}
\arguments{
\item{fcd}{flowframe object (condor)}

\item{old_names}{vector of names that should be changed}

\item{new_names}{vector of new names in the same order}
}
\value{
change_param
}
\description{
change parameter names
}
\details{
change_param_name
}
