% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_IDs}
\alias{check_IDs}
\title{check_IDs}
\usage{
check_IDs(fcdataset)
}
\arguments{
\item{fcdataset}{Flow cytometry dataset to be checked.}
}
\value{
check the integrity and the correctness of a flow cytometry dataset
}
\description{
Check the integrity and the correctness of a flow cytometry dataset
}
\details{
check_IDs
}
