% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{clr}
\alias{clr}
\title{crt_transform}
\usage{
clr(x)
}
\arguments{
\item{x}{matrix to transform}
}
\value{
crt_transform
}
\description{
Data transformation, this function run within the prep_fcd wrapper.
}
\details{
crt_transform
}
