% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_GPU}
\alias{cluster_GPU}
\title{GPU assisted clustering}
\usage{
cluster_GPU(
  fcd,
  algorithm = "leiden",
  seed = 91,
  rapids_dir = "/home/rapids/virtualenv/rapids_singlecell/",
  n_neighbor = 15,
  data_slot = "orig",
  n_pc = ncol(fcd$expr$orig),
  prefix = NULL,
  res = 0.6
)
}
\arguments{
\item{fcd}{flow cytometry data set, that has been subjected to clustering with cyCONDOR.}

\item{algorithm}{which algorithm to use for clustering. Possible choices are leiden or louvain.  Default is leiden.}

\item{seed}{A seed is set for reproducibility.}

\item{rapids_dir}{directory containing the virtual environment for rapids singlecell. Default is directory found in our docker image supporting GPU.}

\item{n_neighbor}{Number of neighbors to use for generating neighbor graph, which is required for the clustering algorithm}

\item{data_slot}{name of the PCA data slot to use for harmonization. If no prefix was added the, \code{orig}.}

\item{n_pc}{Number of PCs to use for neighbor graph generation. If nothing specified, will use all markers in dataset.}

\item{prefix}{Prefix for the output.}

\item{res}{Resolution to use for clustering. Default is 0.6.}
}
\value{
The function returns a fcd with a column in the metadata called either "leiden" or "louvain".
}
\description{
Use GPU assisted Louvain or Leiden algorithm for graph-based clustering.
}
\details{
cluster_GPU


See [Korunsky et al., 2019](https://doi.org/10.1038/s41592-019-0619-0) for more details on the Harmony algorithm.
}
