% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{condor_id}
\alias{condor_id}
\title{condor_ind}
\usage{
condor_id(fcd, id = "")
}
\arguments{
\item{fcd}{flow cytometry dataset}

\item{id}{ID to asign to the condor object}
}
\value{
Prints to the console the detail of the condor object
}
\description{
This function is used to asign and ID to a condor object, if the `id` argument is not used the function will print the current asigned id
}
\details{
condor_id
}
