% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{condor_info}
\alias{condor_info}
\title{condor_info}
\usage{
condor_info(fcd = condor)
}
\arguments{
\item{fcd}{flow cytometry dataset}
}
\value{
Prints to the console the detail of the condor object
}
\description{
This function prints an overview of the details of the condor object
}
\details{
condor_info
}
