% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{condor_session_info}
\alias{condor_session_info}
\title{condor_session_info}
\usage{
condor_session_info(fcd = condor)
}
\arguments{
\item{fcd}{flow cytometry dataset}
}
\value{
fcd with sessionInfo sztored under extras
}
\description{
This function saves the session infos to the fcd object
}
\details{
condor_session_info
}
