% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confusion_HM}
\alias{confusion_HM}
\title{confusion_HM}
\usage{
confusion_HM(
  variables,
  group,
  size = 15,
  title = "Choose a good title for the plot"
)
}
\arguments{
\item{variables}{Variable to by used to calculate the confusion.}

\item{group}{Groupping to calculate the relative contribution to the variable.}

\item{size}{Size of the heatmap blocks.}

\item{title}{Title of the plot.}
}
\value{
confusion_HM
}
\description{
Visualize the confusion heatmap.
}
\details{
confusion_HM
}
