% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{confusionMatrix}
\usage{
confusionMatrix(i = NULL, j = NULL)
}
\arguments{
\item{i}{i}

\item{j}{j}
}
\value{
Create a confusion matrix
}
\description{
Calculate a confusion matrix
}
\details{
confusionMatrix
}
\references{
code for this function was obtained from the R package ArchR (https://github.com/GreenleafLab/ArchR/blob/master/R/HelperUtils.R)
}
