% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_metaclustering_script}
\alias{create_metaclustering_script}
\title{create_metaclustering_script}
\usage{
create_metaclustering_script(num_clusters)
}
\arguments{
\item{num_clusters}{(Numeric) number of clusters to assign.}
}
\value{
create_metaclustering_script
}
\description{
This function created an empty script to assign metaclusters. Thanks to @Lucas for providing this helpful peace of code!
}
\details{
create_metaclustering_script
}
