% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{densityplot_marker}
\alias{densityplot_marker}
\title{dotplot_cyto}
\usage{
densityplot_marker(
  data,
  marker,
  split_by = NULL,
  color_by = NULL,
  color_discrete = cluster_palette,
  title
)
}
\arguments{
\item{data}{Data to use for the plot.}

\item{marker}{Marker to be visualized.}

\item{split_by}{Variable used to split the plot.}

\item{color_by}{Variable used to color the plot.}

\item{color_discrete}{Color palette (default: cluster_palette)}

\item{title}{Plot title.}
}
\value{
densityplot marker.
}
\description{
Plot the distribution of the expression of selected markers, similar to FlowJo Histogram.
}
\details{
densityplot_marker
}
