% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{df_frequency}
\alias{df_frequency}
\title{df_frequency}
\usage{
df_frequency(
  classification,
  classification_header = "classification",
  condition,
  vertical = TRUE,
  groups = NULL
)
}
\arguments{
\item{classification}{classification parameters.}

\item{classification_header}{**optional** chosen header for classification parameters, default = "classification".}

\item{condition}{grouping to be used.}

\item{vertical}{logical, if FALSE frequency on level of classification, default = TRUE.}

\item{groups}{** optional ** vector of selected groups to display, default = all.}
}
\value{
df_frequency
}
\description{
function to get dataframe of frequencies
}
\details{
df_frequency
}
