% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{dotplot_cyto}
\alias{dotplot_cyto}
\title{dotplot_cyto}
\usage{
dotplot_cyto(
  data,
  subset = FALSE,
  annotation,
  color_by,
  subset_char,
  color_discrete = cluster_palette,
  x,
  y,
  title
)
}
\arguments{
\item{data}{data to use for the plot.}

\item{subset}{LOGICLE, if TRUE the data will be subsetted for selected population(s).}

\item{annotation}{Vector of the same lenght at data containing the variable to subset.}

\item{color_by}{Parameter to be used to define the plot colors.}

\item{subset_char}{Vector of variables to keep in the plot (e.g. c("CD4T, "CD8T")).}

\item{color_discrete}{Color palette to use (default: cluster palette).}

\item{x}{Parameter to display in the X axes.}

\item{y}{Parameter to display in the Y axes.}

\item{title}{Plot title.}
}
\value{
2D dotplot.
}
\description{
Plot a classical dotplot such as normally seen in FC analysis (e.g. FlowJo).
}
\details{
dotplot_cyto
}
