% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_fcd}
\alias{filter_fcd}
\title{filter_fcd}
\usage{
filter_fcd(fcdataset, cell_ids)
}
\arguments{
\item{fcdataset}{Flow cytometry dataset to be filtered.}

\item{cell_ids}{Row names of the cells to be filtered, should be provided as vector.}
}
\value{
filter the flowframe according to specific cell IDs
}
\description{
Filter a fcd according to selected cell IDs
}
\details{
filter_fcd
}
