% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmonize_PCA}
\alias{harmonize_PCA}
\title{harmonize_PCA}
\usage{
harmonize_PCA(fcd, data_slot = "orig", batch, seed, prefix = NULL)
}
\arguments{
\item{fcd}{flow cytometry dataset}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{batch}{vector of column names to use for correcting the data}

\item{seed}{Seed used for the randomization steps}

\item{prefix}{Prefix for the output.}
}
\value{
harmonize_PCA
}
\description{
Harmonize the Principal Component Analysis.
}
\details{
harmonize_PCA
}
