% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{HM_differential_marker}
\alias{HM_differential_marker}
\title{HM_differential_marker}
\usage{
HM_differential_marker(
  fcd,
  data_slot = "orig",
  cluster_method,
  cluster_type,
  maxvalue = NULL,
  group_by,
  size = 10,
  title = "Choose a good title for the plot",
  exclusion = c("SSC-A", "FSC-A")
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{data_slot}{data to use for the calculation of the UMAP, e.g. "expr" or "pca".}

\item{cluster_method}{Name of the clustering result to be used.}

\item{cluster_type}{Type of the clustering.}

\item{maxvalue}{Max value for the coloring (Default: NULL, automatically defined).}

\item{group_by}{Grouping variable for visualization.}

\item{size}{Size of the individual squares.}

\item{title}{Title for the plot.}

\item{exclusion}{Vector with markers to exclude from the plot.}
}
\value{
df_frequency
}
\description{
Calculate HM to compare marker expression in different groups.
}
\details{
HM_differential_marker
}
