% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{HM_markers}
\alias{HM_markers}
\title{HM_markers}
\usage{
HM_markers(
  input,
  group,
  maxvalue = NULL,
  size = 10,
  title = "Choose a good title for the plot",
  exclusion = c("SSC-A", "FSC-A"),
  cluster_rows = FALSE,
  cluster_cols = FALSE
)
}
\arguments{
\item{input}{cbind of the expression table to be used the the groupping (eg. clustering).}

\item{group}{column name to be used for the groupping (eg. "Phenograph" or "group")}

\item{maxvalue}{-Max scaled expression to be used for the color coding.}

\item{size}{Size of the cells of the heatmap.}

\item{title}{Title for the plot.}

\item{exclusion}{Marker to exclude from the visualization.}

\item{cluster_rows}{Logical: If heatmap rows should be clustered.}

\item{cluster_cols}{Logical: If heatmap columns should be clustered.}
}
\value{
HM_markers
}
\description{
Visualize the heatmap of marker expression groupped by a variable of interest.
}
\details{
HM_markers
}
