% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_projection.R
\name{learnUMAP}
\alias{learnUMAP}
\title{learnUMAP}
\usage{
learnUMAP(
  fcd,
  input_type,
  data_slot,
  model,
  n_epochs = 100,
  prefix = NULL,
  n_threads = 32,
  seed
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "expr" or "pca".}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{model}{Data associated with an existing embedding.}

\item{n_epochs}{Number of epochs to use during the optimization of the embedded coordinates. A value between 30 - 100 is a reasonable trade off between speed and thoroughness. By default, this value is set to one third the number of epochs used to build the model.}

\item{prefix}{Prefix for the name of the dimensionality reduction.}

\item{n_threads}{Number of threads to use, (except during stochastic gradient descent). Default is half the number of concurrent threads supported by the system.}

\item{seed}{Seed to be used.}
}
\value{
learnUMAP
}
\description{
Uses the model calculated with *runUMAP* to project new samples
}
\details{
learnUMAP
}
