% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{measured_markers}
\alias{measured_markers}
\title{measured_markers}
\usage{
measured_markers(fcd)
}
\arguments{
\item{fcd}{flow cytometry dataset.}
}
\value{
The function returns a vector containing all markers present in the fcd.
}
\description{
returns a vector of markers for fcd$expr$orig.
}
\details{
measured_markers
}
