% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{metaclustering}
\alias{metaclustering}
\title{metaclustering}
\usage{
metaclustering(
  fcd,
  clustering,
  cluster_slot,
  cluster_var = "metacluster",
  metaclusters
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{clustering}{Name of the clustering to match for the metaclustering.}

\item{cluster_slot}{Column containing the original cluster.}

\item{cluster_var}{Name of the output column.}

\item{metaclusters}{Vector of the new clusters names, this should be of the same length of the levels of the original clustering.}
}
\value{
metaclustering
}
\description{
Assignment of a metaclusters name.
}
\details{
metaclustering
}
