% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{nfTransform}
\alias{nfTransform}
\title{nfTransform}
\usage{
nfTransform(transTypeTable, dataA, dataB)
}
\arguments{
\item{transTypeTable}{Table with the transformation parameters}

\item{dataA}{dataA}

\item{dataB}{dataB}
}
\value{
transformed flow cytometry dataset
}
\description{
Data transformation.
}
\details{
nfTransform
}
