% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{order_param}
\alias{order_param}
\title{order_param}
\usage{
order_param(fcd, type, cluster_slot, param, new_order)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{type}{This function can be used to re-order either a metadata annotation or a clustering label (metacluster).}

\item{cluster_slot}{string specifying which clustering slot to use to find variable specified in param. Necessary only if type is "cluster".}

\item{param}{Parameter to reorder.}

\item{new_order}{A vector describing the new order asigned to the labels.}
}
\value{
Returns a condor object with a reordered variable
}
\description{
This function helps to re-oder the levels of a variable to allow a personalized order for plotting.
}
\details{
order_param
}
