% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{PC_loadings}
\alias{PC_loadings}
\title{PC_loadings}
\usage{
PC_loadings(fcd, prefix = NULL, data_slot = "orig", nPC = 3)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{prefix}{Prefix for the output.}

\item{data_slot}{data slot to use for the calculation of the PCA, e.g. "orig" or "norm".}

\item{nPC}{Number of principle components to show.}
}
\value{
Figure of PC loadings
}
\description{
Function to visualize the effect each marker has on the principle component.
}
\details{
PC_loadings
}
