% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_marker}
\alias{plot_marker}
\title{plot_marker}
\usage{
plot_marker(
  data,
  param,
  order = FALSE,
  title = "adjust the title",
  limX = NULL,
  limY = NULL,
  dim_red,
  dot_size = 0.1,
  apha = 0.2,
  color_discrete = cluster_palette,
  color_gradient = colors,
  remove_guide = FALSE,
  facet_by_variable = FALSE,
  label_clusters = FALSE,
  label_size = 3.5,
  label_color = "black"
)
}
\arguments{
\item{data}{This is the input dataframe for the visualization, this part of the code can still be improved but for the moment you have to cbind the dataframe with the informations you want to plot.}

\item{param}{Parameter to visualize in the plot, this can be either a continous variable or a categorical one, the function will react differently according.}

\item{order}{Logical if you want to order the dots in the plot, by expression for example. This can help to find small populations of positive cells.}

\item{title}{Title of the plot.}

\item{limX}{Limits of the x axes.}

\item{limY}{Limits of the y axes.}

\item{dim_red}{Dimensionality reduction to use for the visualization.}

\item{dot_size}{Size of the dots.}

\item{apha}{Alpha of the dots.}

\item{color_discrete}{Colors for discrete parameters.}

\item{color_gradient}{Colors for continous parameters.}

\item{remove_guide}{Logical, if you want to remove the guide.}

\item{facet_by_variable}{Logical if the plot should be splitted by the categorical variable used.}

\item{label_clusters}{XX}

\item{label_size}{XX}

\item{label_color}{XX}
}
\value{
plot marker or list of markers
}
\description{
Function to visualize a selected parameter.
}
\details{
plot_marker
}
