% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_projection.R
\name{predict_labels}
\alias{predict_labels}
\title{predict_labels}
\usage{
predict_labels(
  fcd,
  input_type,
  data_slot,
  fcd_model,
  label = "predicted_labels",
  seed = 91
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "expr" or "pca".}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{fcd_model}{flow cytometry dataset containing Caret model for the label transfer.}

\item{label}{Label for the output column of the condor object.}

\item{seed}{A seed is set for reproducibility.}
}
\value{
predict_labels
}
\description{
Uses the model generated with \code{\link{train_transfer_model}} to predict the labels of new samples
}
\details{
predict_labels
}
