% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{prep_fjw}
\alias{prep_fjw}
\title{Read FlowJo Workspace}
\usage{
prep_fjw(
  data_gs,
  pop = "root",
  gate_list = nodelist,
  inverse.transform = FALSE,
  transformation,
  remove_param,
  merge_anno = FALSE,
  anno_table,
  separator_anno
)
}
\arguments{
\item{data_gs}{Gate Set object from flowWorkspace Package.}

\item{pop}{Gate to keep for downstream analysis (default: 'root').}

\item{gate_list}{Gate List of the FlowJo Workspace.}

\item{inverse.transform}{Logical: if the data should be reverse transformed of kept with FlowJo transformation (default = FALSE).}

\item{transformation}{If inverse.transform = TRUE, type of new transformation to perform (see nfTransform).}

\item{remove_param}{Parameters to be removed from the condor object.}

\item{merge_anno}{Logical: If sample anno should be merged to the condor object.}

\item{anno_table}{Path to annotation table.}

\item{separator_anno}{Separator of the .csv annotation table.}
}
\value{
read_flowjo_workspace
}
\description{
read_flowjo_workspace and prepare the condor object
}
\details{
Read FlowJo workspace
}
