% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astir.R
\name{runAstir_celltype}
\alias{runAstir_celltype}
\title{runAstir_celltype}
\usage{
runAstir_celltype(
  fcd,
  data_slot,
  analysis_path,
  manifest_name,
  max_epochs,
  learning_rate,
  initial_epochs
)
}
\arguments{
\item{fcd}{Flow cytometry dataset.}

\item{data_slot}{Data slot to use for the analysis (e.g. "orig" or "norm").}

\item{analysis_path}{Full path to the output folder of astir analysis.}

\item{manifest_name}{Filename of the manifest file, this file must be located in the `analysis_path` folder.}

\item{max_epochs}{Maximum number of epochs.}

\item{learning_rate}{Learning Rate.}

\item{initial_epochs}{Initial Epochs.}
}
\value{
runAstir_celltype
}
\description{
Predict cell types using Astir. This package requires the python library `astir` and `reticulate` to work.
}
\details{
runAstir_celltype
}
