% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(
  fcd,
  data_slot = "orig",
  seed = 91,
  prefix = NULL,
  markers = colnames(fcd$expr[["orig"]]),
  discard = FALSE
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{data_slot}{data slot to use for the calculation, e.g. "orig" or "norm".}

\item{seed}{A seed is set for reproducibility.}

\item{prefix}{Prefix of the output.}

\item{markers}{vector of marker names to include or exclude from the calculation according to the discard parameter. See functions used_markers and measured_markers for the extraction of markers directly from the condor object}

\item{discard}{LOGICAL if the markers specified should be included, "F", or excluded, "T", from the calculation. Default = F.}
}
\value{
runPCA
}
\description{
Run a Principal Component Analysis.
}
\details{
runPCA
}
