% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runPCA_pseudobulk}
\alias{runPCA_pseudobulk}
\title{runPCA_pseudobulk}
\usage{
runPCA_pseudobulk(fcd)
}
\arguments{
\item{fcd}{flow cytometry dataset.}
}
\value{
runPCA_pseudobulk
}
\description{
run a Pricipal component analysis on pseudobulk samples
}
\details{
runPCA_pseudobulk
}
