% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runPseudotime}
\alias{runPseudotime}
\title{runPseudotime}
\usage{
runPseudotime(subset_fcd, cluster_method, pca_data_slot = "norm")
}
\arguments{
\item{subset_fcd}{XX}

\item{pca_data_slot}{XX}

\item{cluster_methods}{XX}
}
\value{
df_frequency
}
\description{
Calculate speudotime or flow data
}
\details{
runPseudotime
}
