% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runtSNE}
\alias{runtSNE}
\title{runtSNE}
\usage{
runtSNE(
  fcd,
  input_type,
  data_slot,
  perplexity,
  seed = 91,
  prefix = NULL,
  nThreads = 1,
  nPC = ncol(fcd$pca[[data_slot]]),
  markers = colnames(fcd$expr[[data_slot]]),
  discard = FALSE
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation, e.g. "expr" or "pca" (suggested: "pca").}

\item{data_slot}{data slot to use for the calculation, e.g. "orig" or "norm".}

\item{perplexity}{Perplexity used for tSNE calculation (see Rtsne documentation for details).}

\item{seed}{A seed is set for reproducibility.}

\item{prefix}{Prefix of the output.}

\item{nThreads}{Number of threads to be used in the tSNE calculation.}

\item{nPC}{Number of principal components to use for the analysis.}

\item{markers}{vector of marker names to include or exclude from UMAP calculation according to the discard parameter. See functions used_markers and measured_markers for the extraction of markers directly from the condor object}

\item{discard}{Boolean to decide if the markers specified should be included, "F", or excluded, "T", from the UMAP calculation. Default = F.}
}
\value{
tSNE cohordinates
}
\description{
Calculate tSNE dimensionality reduction.
}
\details{
runtSNE
}
