% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scaleColors}
\alias{scaleColors}
\title{scaleColors}
\usage{
scaleColors(data = input_scale, maxvalue = NULL)
}
\arguments{
\item{data}{data to use}

\item{maxvalue}{value at which the color is fully red / blue}
}
\value{
Defines the color coding for an heatmap.
}
\description{
Defines the color coding for an heatmap.
}
\details{
scaleColors
}
