% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_fcd}
\alias{subset_fcd}
\title{subset_fcd}
\usage{
subset_fcd(fcd, size, seed = 91)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{size}{Numeric: size of the sub-sampling.}

\item{seed}{A seed is set for reproducibility.}
}
\value{
df_frequency
}
\description{
Performs a random subset of the fcd
}
\details{
subset_fcd
}
