% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_fcd_byparam}
\alias{subset_fcd_byparam}
\title{subset_fcd_byparam}
\usage{
subset_fcd_byparam(fcd, param, size, seed = 91)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{param}{Name of the parameter to be used to equaly subset the fcd. This should be a columns in the cell annotation table.}

\item{size}{Numeric: size of the sub-sampling for each element in `param`.}

\item{seed}{A seed is set for reproducibility.}
}
\value{
subset_fcd_byparam
}
\description{
Performs a random subset of the fcd
}
\details{
subset_fcd_byparam
}
