% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(
  keep,
  transformation,
  original_data,
  verbose,
  cofactor = 5,
  use_transformation_table,
  transformation_table,
  separator_transformation
)
}
\arguments{
\item{keep}{Vector of the parameter to keep in the analysis.}

\item{transformation}{transformation to perform.}

\item{original_data}{Original data}

\item{verbose}{Logical, if TRUE the transformation parameters are printed.}

\item{cofactor}{cofactor used for 'arcsinh' transformation, default 5, can be set to 150 for HDFC data.}

\item{use_transformation_table}{Default FALSE. If TRUE a transformation table will be used to define parameter specific transformations.}

\item{transformation_table}{Path to the transformation table. For details on the structure see the online documentation.}

\item{separator_transformation}{Separator for the provided transformation_table. Default ",".}
}
\value{
transformed flow cytometry dataset
}
\description{
Data transformation, this function runs within the \code{\link{prep_fcd}} wrapper, the logicle tranformation are derived from Cytofkit.
}
\details{
transform_data
}
