% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{validate_max_cell}
\alias{validate_max_cell}
\title{Function to check the validity of the max_cell input.}
\usage{
validate_max_cell(max_cell, n_files, anno_table)
}
\arguments{
\item{max_cell}{The maximum number of cells to analyse for each file. Passed from 'prep_fcs'.}

\item{n_files}{number of files to read. Passed from 'prep_fcs'.}

\item{anno_table}{Path of the annotaton file. Passed from 'prep_fcs'.}
}
\value{
validate_max_cell
}
\description{
Checks if max_cell is either an integer length == 1 or a numeric vector with length == n_files. If a string, checks for a column in anno_table that matches the number of files.
}
