% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{violinplot_marker}
\alias{violinplot_marker}
\title{violinplot_marker}
\usage{
violinplot_marker(
  fcd,
  data_slot = "orig",
  cluster_method,
  cluster_type,
  marker = colnames(df[, -which(names(df) \%in\% c("metaclusters", "Description",
    "Phenograph", "FlowSOM"))])
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{data_slot}{data to use for the calculation, e.g. "expr" or "pca".}

\item{cluster_method}{methods for the clustering to use.}

\item{cluster_type}{type of clustering (slot name).}

\item{marker}{marker to plot.}
}
\value{
violinplot_marker
}
\description{
Calculate a violin plot for the expression of selected markers.
}
\details{
violinplot_marker
}
