% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{barplot_frequency}
\alias{barplot_frequency}
\title{barplot_frequency}
\usage{
barplot_frequency(
  x_axes,
  colour,
  color_palette = cluster_palette,
  title = "Choose an appropriate title",
  legend_title
)
}
\arguments{
\item{x_axes}{Groupping of the x axes.}

\item{colour}{Stratification to use on the stacked barplot.}

\item{color_palette}{color_palette}

\item{title}{Title for the plot.}

\item{legend_title}{Title for the legend.}
}
\value{
barplot_frequency
}
\description{
The function get_expr_huva provides the expression table resulting from reference huva experiment data.frame.
}
\details{
barplot_frequency
}
