% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{boxplot_and_stats}
\alias{boxplot_and_stats}
\title{boxplot_and_stats}
\usage{
boxplot_and_stats(
  annotation,
  sample_var,
  group_var,
  which_groups = NULL,
  variable,
  numeric = TRUE,
  test.type,
  paired.test = FALSE
)
}
\arguments{
\item{annotation}{Sample annotation to be used for the plot.}

\item{sample_var}{Column name containing the sample IDs.}

\item{group_var}{Column name defining the groupping for plotting.}

\item{which_groups}{XX}

\item{variable}{Variable used to stratify the plotting.}

\item{numeric}{Logical if the groupping is numeric.}

\item{test.type}{Test to be performed. (see need some development here)}

\item{paired.test}{XX}
}
\value{
boxplot_and_stats
}
\description{
Calculate boxplot and summary statistics for a wanted comparison.
}
\details{
boxplot_and_stats
}
