% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{change_param_name}
\alias{change_param_name}
\title{change_param}
\usage{
change_param_name(fcd, old_names, new_names)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{old_names}{vector of names that should be changed.}

\item{new_names}{vector of new names in the same order.}
}
\value{
The function returns a fcd with changed parameter names.
}
\description{
change parameter names
}
\details{
change_param_name
}
