% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_IDs}
\alias{check_IDs}
\title{check_IDs}
\usage{
check_IDs(fcd)
}
\arguments{
\item{fcd}{Flow cytometry dataset to be checked.}
}
\value{
check the integrity and the correctness of a flow cytometry dataset
}
\description{
Check the integrity and the correctness of a flow cytometry dataset by comparing the cell IDs of all existing slots to the cell IDs of the orig. exprs
}
\details{
check_IDs
}
