% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{condor-onAttach}
\alias{condor-onAttach}
\alias{.onAttach}
\title{Print Welcome Message and ASCII Art upon Loading this function is inspired by the developers of the `ggvolc` package}
\usage{
.onAttach(libname, pkgname)
}
\description{
This function is executed when the `cyCONDOR` package is attached to the R session.
It prints a welcome message and an ASCII representation related to the package.
}
\keyword{internal}
