% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{confusion_HM}
\alias{confusion_HM}
\title{confusion_HM}
\usage{
confusion_HM(
  variables,
  group,
  size = 15,
  title = "Choose a good title for the plot",
  cluster_cols = FALSE,
  cluster_rows = FALSE
)
}
\arguments{
\item{variables}{Variable to by used to calculate the confusion.}

\item{group}{Groupping to calculate the relative contribution to the variable.}

\item{size}{Size of the heatmap blocks.}

\item{title}{Title of the plot.}

\item{cluster_cols}{TRUE or FALSE if columns should be clustered}

\item{cluster_rows}{TRUE or FALSE if rows should be clustered}
}
\value{
confusion_HM
}
\description{
Visualize the confusion heatmap.
}
\details{
confusion_HM
}
