% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{confusionMatrix}
\usage{
confusionMatrix(i = NULL, j = NULL)
}
\arguments{
\item{i}{i}

\item{j}{j}
}
\value{
Create a confusion matrix
}
\description{
Calculate a confusion matrix
}
\details{
confusionMatrix
}
