% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{filter_fcd}
\alias{filter_fcd}
\title{filter_fcd}
\usage{
filter_fcd(fcd, cell_ids)
}
\arguments{
\item{fcd}{Flow cytometry dataset to be filtered.}

\item{cell_ids}{Row names of the cells to be filtered, should be provided as vector.}
}
\value{
The function returns a fcd filtered on the specified cell IDs.
}
\description{
Filters a fcd according to selected cell IDs
}
\details{
filter_fcd
}
