% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{harmonize_intensities}
\alias{harmonize_intensities}
\title{harmonize_intensities}
\usage{
harmonize_intensities(fcd, batch_var, seed = 91)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{batch_var}{vector of column names from \code{fcd$anno$cell_anno} to use for correcting the data.}

\item{seed}{A seed is set for reproducibility.}
}
\value{
fcd with a harmonized expression data frame.
}
\description{
Harmonize the expression values.
}
\details{
harmonize_intensities


See https://doi.org/10.1038/s41592-019-0619-0 for more details on the Harmony algorithm.
}
