% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{harmonize_intensities}
\alias{harmonize_intensities}
\title{harmonize_intensities}
\usage{
harmonize_intensities(fcd, batch, seed)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{batch}{vector of column names to use for correcting the data.}

\item{seed}{Seed used for the randomization steps.}
}
\value{
harmonize_intensities
}
\description{
Harmonize the expression values.
}
\details{
harmonize_intensities
}
