% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{harmonize_PCA}
\alias{harmonize_PCA}
\title{harmonize_PCA}
\usage{
harmonize_PCA(fcd, data_slot = "orig", batch_var, seed = 91, prefix = NULL)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, \code{orig}.}

\item{batch_var}{vector of column names from \code{fcd$anno$cell_anno} to use for correcting the data.}

\item{seed}{A seed is set for reproducibility.}

\item{prefix}{Prefix for the output.}
}
\value{
harmonize_PCA
}
\description{
Harmonize the Principal Component Analysis.
}
\details{
harmonize_PCA


See https://doi.org/10.1038/s41592-019-0619-0 for more details on the Harmony algorithm.
}
