% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{HM_differential_marker}
\alias{HM_differential_marker}
\title{HM_differential_marker}
\usage{
HM_differential_marker(
  fcd,
  data_slot = "orig",
  cluster_method,
  cluster_type,
  maxvalue = NULL,
  size = 10,
  title = "Choose a good title for the plot",
  exclusion = c("SSC-A", "FSC-A")
)
}
\arguments{
\item{fcd}{XX}

\item{data_slot}{XX}

\item{cluster_method}{XX}

\item{cluster_type}{XX}

\item{maxvalue}{XX}

\item{size}{XX}

\item{title}{XX}

\item{exclusion}{XX}

\item{pca_data_slot}{XX}
}
\value{
df_frequency
}
\description{
Calculate speudotime or flow data
}
\details{
HM_differential_marker
}
