% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_projection.R
\name{learnUMAP}
\alias{learnUMAP}
\title{learnUMAP}
\usage{
learnUMAP(
  fcd,
  input_type,
  data_slot,
  model,
  n_epochs = 100,
  prefix = NULL,
  n_threads = 32,
  seed
)
}
\arguments{
\item{fcd}{XX}

\item{input_type}{XX}

\item{data_slot}{XX}

\item{model}{XX}

\item{n_epochs}{XX}

\item{prefix}{XX}

\item{n_threads}{XX}

\item{seed}{XX}
}
\value{
learnUMAP
}
\description{
learnUMAP
}
\details{
learnUMAP
}
