% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{measured_markers}
\alias{measured_markers}
\title{measured_markers}
\usage{
measured_markers(fcd)
}
\arguments{
\item{fcd}{flow cytometry dataset.}
}
\value{
marker vector.
}
\description{
returns a vector of markers for orig condor expr.
}
\details{
measured_markers
}
