% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_condor}
\alias{merge_condor}
\title{merge_condor}
\usage{
merge_condor(data1, data2)
}
\arguments{
\item{data1}{Dataset 1 to merge}

\item{data2}{Dataset 2 to merge}
}
\value{
Condor Object
}
\description{
Merges two condor objects.
}
\details{
merge_condor
}
