% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{PC_loadings}
\alias{PC_loadings}
\title{PC_loading}
\usage{
PC_loadings(fcd, data_slot = "orig", number)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{data_slot}{data to use for the calculation of the UMAP, e.g. "expr" or "pca".}

\item{number}{Number of principal components to plot.}
}
\value{
PC loadings
}
\description{
Calculate PCA loading.
}
\details{
PC_loading
}
