% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PC_loadings}
\alias{PC_loadings}
\title{PC_loading}
\usage{
PC_loadings(fcd, data_slot = "orig", number)
}
\arguments{
\item{fcd}{XX}

\item{data_slot}{XX}

\item{number}{XX}
}
\value{
PC loadings
}
\description{
Calculate speudotime or flow data
}
\details{
PC_loading
}
