% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(
  data,
  param,
  title = "adjust the title",
  dim_red,
  dot_size = 0.1,
  alpha = 0.2,
  color_density = c("Blues", "Oranges", "Reds")
)
}
\arguments{
\item{data}{XX}

\item{param}{XX}

\item{title}{XX}

\item{dim_red}{XX}

\item{dot_size}{XX}

\item{alpha}{XX}

\item{color_density}{XX}
}
\value{
Density Plot
}
\description{
Density Plot of dimensionality reduction.
}
\details{
plot_density
}
