% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(
  data,
  param,
  title = "adjust the title",
  dim_red,
  dot_size = 0.1,
  alpha = 0.2,
  color_density = c("Blues", "Oranges", "Reds")
)
}
\arguments{
\item{data}{data to be used for plotting.}

\item{param}{Parameter to be used for the density splitting.}

\item{title}{Title of the plot.}

\item{dim_red}{Dimensionality reduction to be used.}

\item{dot_size}{Size of the background dots.}

\item{alpha}{Transparency of the background dots.}

\item{color_density}{Colors of the density maps.}
}
\value{
Density Plot
}
\description{
Density Plot of dimensionality reduction.
}
\details{
plot_density
}
