% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_classifier.R
\name{predict_classifier}
\alias{predict_classifier}
\title{predict_classifier}
\usage{
predict_classifier(
  fcd,
  input_type,
  data_slot,
  sample_names = "expfcs_filename",
  model_object,
  seed = 91
)
}
\arguments{
\item{fcd}{flowframe object.}

\item{input_type}{data to use.}

\item{data_slot}{name of the data slot to use.}

\item{sample_names}{Column name of the metadata table containing the file names.}

\item{model_object}{flowframe object with the stored classifier model.}

\item{seed}{A seed is set for reproducibility.}
}
\value{
predict_classifier
}
\description{
This function uses the model trained with *train_classifier_model* to predict new samples.
}
\details{
Predict Clinical Classifier
}
