% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_classifier.R
\name{predict_classifier}
\alias{predict_classifier}
\title{predict_classifier}
\usage{
predict_classifier(
  fcd,
  input_type,
  data_slot,
  sample_names = "expfcs_filename",
  model_object,
  seed
)
}
\arguments{
\item{fcd}{XX}

\item{input_type}{XX}

\item{data_slot}{XX}

\item{sample_names}{XX}

\item{model_object}{XX}

\item{seed}{XX}
}
\value{
predict_classifier
}
\description{
predict_classifier
}
\details{
Predict Clinical Classifier
}
