% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_labels}
\alias{predict_labels}
\title{predict_labels}
\usage{
predict_labels(
  fcd,
  input_type,
  data_slot,
  model_object,
  label = "predicted_labels",
  seed
)
}
\arguments{
\item{fcd}{XX}

\item{input_type}{XX}

\item{data_slot}{XX}

\item{model_object}{XX}

\item{label}{XX}

\item{seed}{XX}
}
\value{
predict_labels
}
\description{
predict_labels
}
\details{
predict_labels
}
