% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{prep_fcd}
\alias{prep_fcd}
\title{prep_fcd}
\usage{
prep_fcd(
  data_path,
  max_cell,
  useCSV = FALSE,
  transformation,
  remove_param,
  anno_table,
  filename_col,
  seed = 91,
  separator_anno = ",",
  separator_fc_csv = ","
)
}
\arguments{
\item{data_path}{Folder where the .fcs files or .csv files are stored.}

\item{max_cell}{Number of cells to use for each file (set to a high number if you want to use all available events).}

\item{useCSV}{Flag if the input are .csv files and not .fcs (experimental).}

\item{transformation}{Transformation to perform.}

\item{remove_param}{Parameters to remove from the transformation, "inTime" should be kept.}

\item{anno_table}{path to the annotation table file.}

\item{filename_col}{Name of the column containing the file name matching with the .fcs files.}

\item{seed}{A seed is set for reproducibility.}

\item{separator_anno}{separator used in the annotation file.}

\item{separator_fc_csv}{separator used in the fc csv files.}
}
\value{
prep_fcd
}
\description{
Wrapping function to prepare a flow cytometry dataset
}
\details{
prep_fcd
}
