% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prep_fcd}
\alias{prep_fcd}
\title{prep_fcd}
\usage{
prep_fcd(
  FCSpath,
  ceil,
  useCSV = FALSE,
  transformation,
  remove_param,
  anno_table,
  filename_col,
  seed
)
}
\arguments{
\item{FCSpath}{Folder where the .fcs files are stored.}

\item{ceil}{Number of cells to use for each file (set to a high number if you want to use all available events)}

\item{useCSV}{Flag if the input are .csv files and not .fcs (experimental).}

\item{transformation}{Transformation to perform.}

\item{remove_param}{Parameters to remove from the trasfomration, "inTime" should be kept.}

\item{anno_table}{path to the annotation table file.}

\item{filename_col}{Name of the column containing the filename matching with the .fcs files.}

\item{seed}{seed to be used for the randomization of the events.}
}
\value{
prep_fcd
}
\description{
Wrapping function to prepare a flow cytometry dataset
}
\details{
prep_fcd
}
