% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{prep_fjw}
\alias{prep_fjw}
\title{read_flowjo_workspace}
\usage{
prep_fjw(
  data_gs,
  pop = "root",
  gate_list = nodelist,
  inverse.transform = FALSE,
  transformation = "a",
  remove_param,
  merge_anno = FALSE,
  anno_table,
  separator_anno
)
}
\arguments{
\item{data_gs}{XX}

\item{pop}{XX}

\item{gate_list}{XX}

\item{inverse.transform}{XX}

\item{transformation}{XX}

\item{remove_param}{XX}

\item{merge_anno}{XX}

\item{anno_table}{XX}

\item{separator_anno}{XX}
}
\value{
read_flowjo_workspace
}
\description{
read_flowjo_workspace
}
\details{
Read FlowJo workspace
}
