% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load_and_transform.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(FCSpath, max_cells, useCSV, separator)
}
\arguments{
\item{FCSpath}{Path to the .fcs files.}

\item{max_cells}{number of cells to subset.}

\item{useCSV}{Logical, if input is .csv and not .fcs.}

\item{separator}{Separator used the flow csv files (if loading from csv).}
}
\value{
load flow cytometry dataset
}
\description{
Load .fcs or .csv files into a dataframe and prepare the condor object.
}
\details{
read_data
}
