% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runDM}
\alias{runDM}
\title{runDM}
\usage{
runDM(
  fcd,
  input_type,
  data_slot,
  k = 10,
  seed = 91,
  prefix = NULL,
  nPC = ncol(fcd[[input_type]][[data_slot]]),
  markers = colnames(fcd$expr[[data_slot]]),
  discard = FALSE
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation, e.g. "expr" or "pca".}

\item{data_slot}{data slot to use for the calculation, e.g. "orig" or "norm".}

\item{k}{K used for the analysis. Default = 10.}

\item{seed}{A seed is set for reproducibility.}

\item{prefix}{Prefix of the output.}

\item{nPC}{Number of principal components to use for the analysis. Default = All.}

\item{markers}{vector of marker names to include or exclude from DM calculation according to the discard parameter. See functions used_markers and measured_markers for the extraction of markers directly from the condor object}

\item{discard}{LOGICAL if the markers specified should be included, "False", or excluded, "True", from the DM calculation. Default = F.}
}
\value{
runDM
}
\description{
Run Diffusion Map dimensionality reduction.
}
\details{
runDM
}
