% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runDM}
\alias{runDM}
\title{runDM}
\usage{
runDM(
  fcd,
  input_type,
  data_slot,
  k = 10,
  seed,
  prefix = NULL,
  top_PCA = ncol(fcd[[input_type]][[data_slot]])
)
}
\arguments{
\item{fcd}{flow cytometry dataset}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "pca" (suggested option)}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{k}{K used for the analysis.}

\item{seed}{Seed used for the randomization steps}

\item{prefix}{Prefix of the output.}
}
\value{
runDM
}
\description{
Run Diffusion Map dimensionality reduction.
}
\details{
runDM
}
