% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runFlowSOM}
\alias{runFlowSOM}
\title{runFlowSOM}
\usage{
runFlowSOM(fcd, input_type, data_slot, num_clusters, seed, prefix = NULL)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "expr" or "pca".}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{num_clusters}{number of final clusters.}

\item{seed}{Seed used for the randomization steps.}

\item{prefix}{Prefic for the output.}
}
\value{
metaclustering
}
\description{
Run FlowSOM based clustering.
}
\details{
runFlowSOM
}
