% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony.R
\name{runPCA}
\alias{runPCA}
\title{runPCA}
\usage{
runPCA(fcd, data_slot = "orig", seed, prefix = NULL)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{data_slot}{name of the data slot to use to calculate the PCA, original data (orig) or harmonized data (norm).}

\item{seed}{Seed used for the randomization steps.}

\item{prefix}{Prefix for the output.}
}
\value{
runPCA
}
\description{
Run a Principal Component Analysis.
}
\details{
runPCA
}
