% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runPCA_pseudobulk}
\alias{runPCA_pseudobulk}
\title{runPCA_pseudobulk}
\usage{
runPCA_pseudobulk(fcd)
}
\arguments{
\item{fcd}{XX}
}
\value{
runPCA_pseudobulk
}
\description{
runPCA_pseudobulk
}
\details{
runPCA_pseudobulk
}
