% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runPhenograph}
\alias{runPhenograph}
\title{runPhenograph}
\usage{
runPhenograph(fcd, input_type, data_slot, k, seed, prefix = NULL)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "pca" (suggested option).}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{k}{K value used for clustering.}

\item{seed}{Seed used for the randomization steps.}

\item{prefix}{Prefix for the output.}
}
\value{
runPhenograph
}
\description{
Run Phenograph clustering.
}
\details{
runPhenograph
}
