% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{runPseudotime}
\alias{runPseudotime}
\title{runPseudotime}
\usage{
runPseudotime(
  fcd,
  dim_red_type,
  dim_red_name,
  start.clus = NULL,
  end.clus = NULL,
  clustering,
  dist.method = "slingshot",
  use.median = FALSE,
  omega = FALSE,
  omega_scale = 1.5,
  times = NULL,
  shrink = TRUE,
  extend = "y",
  reweight = TRUE,
  reassign = TRUE,
  thresh = 0.001,
  maxit = 15,
  stretch = 2,
  approx_points = NULL,
  smoother = "smooth.spline",
  shrink.method = "cosine",
  allow.breaks = TRUE,
  seed
)
}
\arguments{
\item{fcd}{XX}

\item{dim_red_type}{XX}

\item{dim_red_name}{XX}

\item{start.clus}{XX}

\item{end.clus}{XX}

\item{clustering}{XX}

\item{dist.method}{XX}

\item{use.median}{XX}

\item{omega}{XX}

\item{omega_scale}{XX}

\item{times}{XX}

\item{shrink}{XX}

\item{extend}{XX}

\item{reweight}{XX}

\item{reassign}{XX}

\item{thresh}{XX}

\item{maxit}{XX}

\item{stretch}{XX}

\item{approx_points}{XX}

\item{smoother}{XX}

\item{shrink.method}{XX}

\item{allow.breaks}{XX}

\item{seed}{XX}
}
\value{
runPseudotime
}
\description{
Calculate speudotime or flow data
}
\details{
runPseudotime
}
