% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runtSNE}
\alias{runtSNE}
\title{runtSNE}
\usage{
runtSNE(
  fcd,
  input_type,
  data_slot,
  perplexity,
  seed,
  prefix = NULL,
  n_threads = 1,
  top_PCA = ncol(fcd[[input_type]][[data_slot]])
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation, e.g. "pca" (suggested option).}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{perplexity}{Perplexity used for tSNE calculation (see Rtsne documentation for details).}

\item{seed}{Seed used for the randomization steps.}

\item{prefix}{Prefix of the output.}

\item{n_threads}{Number of threads to be used in the tSNE calculation.}

\item{top_PCA}{Number of principal components to use for the analysis.}
}
\value{
tSNE cohordinates
}
\description{
Calculate tSNE dimensionality reduction..
}
\details{
runtSNE
}
