% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runtSNE}
\alias{runtSNE}
\title{runtSNE}
\usage{
runtSNE(
  fcd,
  input_type,
  data_slot,
  perplexity,
  seed,
  prefix = NULL,
  n_threads = 1,
  top_PCA = ncol(fcd[[input_type]][[data_slot]])
)
}
\arguments{
\item{fcd}{XX}

\item{input_type}{XX}

\item{data_slot}{XX}

\item{perplexity}{XX}

\item{seed}{XX}

\item{prefix}{XX}

\item{n_threads}{XX}

\item{top_PCA}{XX}
}
\value{
tSNE cohordinates
}
\description{
Calvulate tSNE dimensionality reduction
}
\details{
runtSNE
}
