% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{runUMAP}
\alias{runUMAP}
\title{runUMAP}
\usage{
runUMAP(
  fcd,
  input_type,
  data_slot,
  n_neighbors = 15,
  n_components = 2,
  min_dist = 0.2,
  metric = "euclidean",
  seed,
  prefix = NULL
)
}
\arguments{
\item{fcd}{flow cytometry dataset}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "expr" or "pca"}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{n_neighbors}{n_neighbors}

\item{n_components}{n_components}

\item{min_dist}{min_dist}

\item{metric}{metric}

\item{seed}{Seed used for the randomization steps}

\item{prefix}{Prefix for the output.}
}
\value{
runUMAP
}
\description{
Run a UMAP dimensionality reduction.
}
\details{
runUMAP
}
