% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduction.R
\name{runUMAP}
\alias{runUMAP}
\title{runUMAP}
\usage{
runUMAP(
  fcd,
  input_type,
  data_slot,
  n_neighbors = 15,
  n_components = 2,
  min_dist = 0.2,
  metric = "euclidean",
  seed,
  prefix = NULL,
  n_threads = 32,
  top_PCA = ncol(fcd[[input_type]][[data_slot]]),
  ret_model = FALSE
)
}
\arguments{
\item{fcd}{flow cytometry dataset}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "expr" or "pca"}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{n_neighbors}{n_neighbors}

\item{n_components}{n_components}

\item{min_dist}{min_dist}

\item{metric}{metric}

\item{seed}{Seed used for the randomization steps}

\item{prefix}{Prefix for the output.}

\item{n_threads}{Number of threads to be used in the UMAP calculation.}

\item{top_PCA}{Number of PCA used in the UMAP calculation}

\item{ret_model}{LOGICAL if the UMAP model should be saved for future projection of the data}
}
\value{
runUMAP
}
\description{
Run a UMAP dimensionality reduction.
}
\details{
runUMAP
}
