% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_fcd}
\alias{subset_fcd}
\title{subset_fcd}
\usage{
subset_fcd(fcd, size)
}
\arguments{
\item{fcd}{XX}

\item{size}{XX}
}
\value{
df_frequency
}
\description{
Performs a random subset of the fcd
}
\details{
subset_fcd
}
