% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_classifier.R
\name{train_classifier_model}
\alias{train_classifier_model}
\title{train_classifier_model}
\usage{
train_classifier_model(
  fcd,
  input_type,
  data_slot,
  sample_names = "expfcs_filename",
  classification_variable,
  family = "binomial",
  type1 = "response",
  type2 = "response",
  parallelCore = 1,
  reg = FALSE,
  seed
)
}
\arguments{
\item{fcd}{XX}

\item{input_type}{XX}

\item{data_slot}{XX}

\item{sample_names}{XX}

\item{classification_variable}{XX}

\item{family}{XX}

\item{type1}{XX}

\item{type2}{XX}

\item{parallelCore}{XX}

\item{reg}{XX}

\item{seed}{XX}
}
\value{
train_classifier_model
}
\description{
train_classifier_model
}
\details{
Train Clinical Classifier
}
