% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_transfer_model}
\alias{train_transfer_model}
\title{train_transfer_model}
\usage{
train_transfer_model(
  fcd,
  input_type,
  data_slot,
  label,
  method = "knn",
  tuneLength = 5,
  trControl = trainControl(method = "cv"),
  seed
)
}
\arguments{
\item{fcd}{XX}

\item{input_type}{XX}

\item{data_slot}{XX}

\item{label}{XX}

\item{method}{XX}

\item{trControl}{XX}

\item{seed}{XX}

\item{tuneLenght}{XX}
}
\value{
train_transfer_model
}
\description{
train_transfer_model
}
\details{
train_transfer_model
}
