% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_projection.R
\name{train_transfer_model}
\alias{train_transfer_model}
\title{train_transfer_model}
\usage{
train_transfer_model(
  fcd,
  input_type,
  data_slot,
  label,
  method = "knn",
  tuneLength = 5,
  trControl = trainControl(method = "cv"),
  seed
)
}
\arguments{
\item{fcd}{flow cytometry dataset.}

\item{input_type}{data to use for the calculation of the UMAP, e.g. "expr" or "pca".}

\item{data_slot}{name of the PCA data slot to use to harmonize. If no prefix was added the, *orig*.}

\item{label}{Vector with the labels to be used for the label transfer.}

\item{method}{A string specifying which classification or regression model to use. Possible values are found using names(getModelInfo()). See http://topepo.github.io/caret/train-models-by-tag.html. A list of functions can also be passed for a custom model function. See http://topepo.github.io/caret/using-your-own-model-in-train.html for details.}

\item{tuneLength}{An integer denoting the amount of granularity in the tuning parameter grid. By default, this argument is the number of levels for each tuning parameters that should be generated by train. If trainControl has the option search = "random", this is the maximum number of tuning parameter combinations that will be generated by the random search. (NOTE: If given, this argument must be named.)}

\item{trControl}{A list of values that define how this function acts. See trainControl and http://topepo.github.io/caret/using-your-own-model-in-train.html. (NOTE: If given, this argument must be named.)}

\item{seed}{Seed to be used.}
}
\value{
train_transfer_model
}
\description{
Train a machine learning model to transfer cell label (this function implement the *caret* workflow)
}
\details{
train_transfer_model
}
