% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{violinplot_marker}
\alias{violinplot_marker}
\title{violinplot_marker}
\usage{
violinplot_marker(
  fcd,
  data_slot = "orig",
  cluster_method,
  cluster_type,
  marker = colnames(df[, -which(names(df) \%in\% c("metaclusters", "Description",
    "Phenograph", "FlowSOM"))])
)
}
\arguments{
\item{fcd}{XX}

\item{data_slot}{XX}

\item{cluster_method}{XX}

\item{cluster_type}{XX}

\item{marker}{XX}
}
\value{
violinplot_marker
}
\description{
violinplot_marker
}
\details{
violinplot_marker
}
